[path_planning] First compiling draft of Lazy Theta Star node serving requests async.
